refactor(fiscal): decouple fiscal module from database layer
Move invoice-service, invoice-repository and fiscal-settings-repository
out of fiscal/ into lib/, making the fiscal module a pure library with
zero DB/ORM dependencies — ready for extraction as standalone package.
